# docs/openapi.yaml
openapi: 3.0.3
info:
  title: vidlink API
  version: "0.88.0"
  description: |
    多平台视频解析 API：输入分享链接或平台内容 ID，返回平台已发布的直链。
    支持抖音、哔哩哔哩、快手、小红书。人读版本见 API.md，系数见 配额倍率表.md。

    计量单位是**配额**（管理员分配的用量额度，不是余额、与支付无关，API 中无金额字段）：
    `消耗 = 端点系数(端点, 平台) × 条数 × 账号倍率`，只在实际成功时扣减，配额不足返回 429。

    默认只返回 CDN 直链，服务端不代理视频字节。直链通常带 Referer 防盗链，
    浏览器直接 `<video src>` 可能 403：请带上 `Links.headers`，或启用 `/v1/proxy`（默认关闭）。

    鉴权三选一：`X-API-Key` 头 / `Authorization: Bearer` / `?key=`。失败一律 403 且不区分原因。

    ## 公共 Key（vl_public）

    账户模式下服务会准备一个公共账号（`VIDLINK_PUBLIC_KEY`，默认 `vl_public`）：
    Key 公开、免注册，额度是**每 IP 每日 100 配额**（`VIDLINK_PUBLIC_DAILY_QUOTA`），
    与其他端点同一套系数（代理按 1 配额/MiB）。超限返回 `429 public_quota_exhausted`，
    额度按本地时区零点重置。没带 Key 时 `403` 的 message 里会直接给出这个 Key。
    公共 Key 调 `/v1/ledger` 一律 `403`（共享账号，流水混有所有访客）；
    `/v1/usage` 对公共 Key 返回的是"本 IP 今天"的额度（`public: true` + `daily`）。
    每 IP 的日计数只在内存里，重启清零。并发闸门对公共 Key 按 **IP** 生效。

    ## 媒体代理的配额

    `/v1/proxy` 是唯一**按传输体积**计费的出口：`实扣 = 体积(MiB) × 1 × 账号倍率`，
    不乘端点系数、也不乘平台系数（代理搬的是任意 CDN 的字节，与上游解析成本无关）。
    上游声明了长度时先扣后传（不足直接 429，不发出字节），长度未知时边传边限、
    传完按实际字节补扣；HEAD 不计费，提前中断不退。

    ## 管理面（/v1/admin/*）

    管理凭据是**一个固定的 Key**（环境变量或 .vl 里的 `VIDLINK_ADMIN_KEY`），
    不是账本里的账号、也不是账号上的权限位：账号 Key 一律进不了管理面（403），
    管理 Key 也不能用来解析（它不在账本里）。**没有配置这个 Key 时，
    每条 /v1/admin/* 都恒返回 403**（message 里会点名该配置项）。
    账号路径参数 `{key}` 可以是明文 Key，也可以是账号的公开句柄 `id`（`acc_…`，
    Key 的 SHA-256 截断），后者让管理面板不必接触明文 Key。

    `VL_WEBUI=true` 时还提供两个自带页面：`/`（解析页）与 `/admin`（管理面板，
...

...

...
servers:
...
paths:
...
  /v1/ledger:
    get:
      tags: [账号]
      summary: 自己的配额流水（账单）：使用 / 管理员增加 / 减少 / 设置
      description: |
        只能读自己的（服务端按 Key 限定）。**不消耗配额**。
        units 带符号（消耗为负）；totals 为全量口径，entries 只保留最近若干条。
      operationId: getLedger
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: type
          in: query
          schema: { type: string, enum: [consume, add, reduce, set, create, delete] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
        "400": { $ref: "#/components/responses/BadInput" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/ledger:
    get:
      tags: [管理]
      summary: 配额流水：总账单，或指定账号（id 句柄 / key 明文）
      description: 不带 id/key 即总账单。**不消耗配额**。
      operationId: adminLedger
      parameters:
        - name: id
          in: query
          description: 账号公开句柄 acc_…
          schema: { type: string }
        - name: key
          in: query
          description: 账号明文 Key（与 id 二选一）
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: type
          in: query
          schema: { type: string, enum: [consume, add, reduce, set, create, delete] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
        "400": { $ref: "#/components/responses/BadInput" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/accounts:
    get:
      tags: [管理]
      summary: 账号列表（Key 掩码）
      operationId: adminListAccounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts: { type: array, items: { $ref: "#/components/schemas/Account" } }
                  stats: { $ref: "#/components/schemas/AccountStats" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [管理]
      summary: 创建账号
      description: 不填 `key` 时自动生成 256 位随机 Key，明文只在本次响应返回一次。
      operationId: adminCreateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccountRequest" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  account: { $ref: "#/components/schemas/Account" }
                  key: { type: string, description: 自动生成时的明文 Key（只此一次） }
                  notice: { type: string }
        "400": { $ref: "#/components/responses/BadInput" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/accounts/{key}:
    parameters:
      - name: key
        in: path
        required: true
        description: 账号的明文 Key，或列表返回的公开句柄 id（acc_…）
        schema: { type: string }
    get:
      tags: [管理]
      summary: 查看单个账号（Key 掩码）
      operationId: adminGetAccount
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { account: { $ref: "#/components/schemas/Account" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [管理]
      summary: 修改账号（配额 / 倍率 / 停用）
      operationId: adminPatchAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchAccountRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { account: { $ref: "#/components/schemas/Account" } }
        "400": { $ref: "#/components/responses/BadInput" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [管理]
      summary: 删除账号（管理 Key 不在账本里，删谁都不影响管理面）
      operationId: adminDeleteAccount
      responses:
        "200":
          description: 已删除
          content:
            application/json:
              schema:
                type: object
                properties: { deleted: { type: boolean, example: true } }
        "400": { $ref: "#/components/responses/BadInput" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/stats:
    get:
      tags: [管理]
      summary: 运行统计（账号 / 闸门 / 缓存 / 流量）
      operationId: adminStats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts: { $ref: "#/components/schemas/AccountStats" }
                  gate: { $ref: "#/components/schemas/GateStats" }
                  cache: { $ref: "#/components/schemas/CacheStats" }
                  traffic:
                    type: object
                    properties:
                      requests: { type: integer, format: int64 }
                      errors: { type: integer, format: int64 }
                      uptime_sec: { type: integer }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/quota:
    get:
      tags: [管理]
      summary: 配额系数全貌（只读）
      description: 系数是共享参数，影响所有账号，因此只提供读接口；单账号调整请改 `multiplier`。
      operationId: adminQuota
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/QuotaTable" } } }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    ApiKeyHeader: { type: apiKey, in: header, name: X-API-Key }
    BearerAuth: { type: http, scheme: bearer }
    ApiKeyQuery: { type: apiKey, in: query, name: key }
...
  responses:
    BadInput:
      description: 参数不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 缺少或无效 Key、账号停用、需要登录态
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 内容不存在或已删除
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [kind, message]
          properties:
            kind:
              type: string
              enum: [bad_input, unsupported, not_found, forbidden, rate_limited,
                     concurrency_limited, quota_exhausted, timeout, upstream_error,
                     unavailable, method_not_allowed, internal]
            message: { type: string }
            request_id: { type: string }
...
    CacheStats:
      type: object
      description: coalesced 是被请求合并省掉的上游调用数
      properties:
        entries: { type: integer }
        hits: { type: integer, format: int64 }
        misses: { type: integer, format: int64 }
        coalesced: { type: integer, format: int64 }
        loads: { type: integer, format: int64 }
...
    Account:
      type: object
      description: 账号**没有任何权限位**：管理权限属于固定的管理 Key，不属于账号
      properties:
        id: { type: string, description: "公开句柄 acc_…（Key 的 SHA-256 截断，可用于路径参数）" }
        key: { type: string, description: 掩码后的 Key（仅创建时返回明文） }
        name: { type: string }
        note: { type: string }
        quota: { type: number, description: 剩余可用配额 }
        used: { type: number }
        calls: { type: integer }
        multiplier: { type: number }
        disabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Ledger:
      type: object
      description: 配额流水（账单）。不消耗配额；totals 为全量口径
      properties:
        scope: { type: string, enum: [self, all, account] }
        account:
          type: object
          additionalProperties: true
          description: scope=self/account 时带上该账号的摘要
        unit: { type: string, example: 配额 }
        entries:
          type: array
          items: { $ref: "#/components/schemas/LedgerEntry" }
        totals:
          type: object
          additionalProperties:
            type: object
            properties:
              count: { type: integer, format: int64 }
              units: { type: number, description: 带符号净变化；累计消耗 = -totals.consume.units }
        types:
          type: object
          additionalProperties: { type: string }
          description: 各类型的中文含义（由服务端给出，避免前端各写一份）
        limit_max: { type: integer }
        note: { type: string }
    LedgerEntry:
      type: object
      properties:
        time: { type: string, format: date-time }
        type: { type: string, enum: [consume, add, reduce, set, create, delete] }
        key: { type: string, description: 掩码后的 Key }
        id: { type: string, description: 账号公开句柄 }
        name: { type: string }
        units: { type: number, description: 带符号变化量：消耗为负、管理员增加为正 }
        balance: { type: number, description: 这条流水之后的剩余配额 }
        used: { type: number }
        calls: { type: integer, format: int64 }
        detail: { type: string, description: "人能读的说明，例如 links/bilibili ×1、proxy 66.53 MiB（按体积）" }
    AccountStats:
      type: object
      properties:
        accounts: { type: integer }
        active: { type: integer }
        used_total: { type: number }
        calls_total: { type: integer, format: int64 }
        quota_total: { type: number }
    GateStats:
      type: object
      properties:
        global_limit: { type: integer }
        global_in_use: { type: integer }
        global_free: { type: integer }
        waiting: { type: integer, format: int64 }
        max_queue: { type: integer }
        active_keys: { type: integer }
        rejected_key_busy: { type: integer, format: int64 }
        rejected_queue_full: { type: integer, format: int64 }
        timed_out: { type: integer, format: int64 }
    CreateAccountRequest:
      type: object
      properties:
        key: { type: string, description: 留空则自动生成 256 位随机 Key }
        name: { type: string }
        quota: { type: number, description: 初始配额 }
        multiplier: { type: number, default: 1, description: "缺省 1.0；0 表示不扣配额，是有含义的取值" }
        note: { type: string }
    PatchAccountRequest:
      type: object
      description: 全部字段可选；quota 与 add_quota 不能同时使用
      properties:
        name: { type: string }
        note: { type: string }
        quota: { type: number, description: 设为该值（剩余配额） }
        add_quota: { type: number, description: 在现有剩余配额上加减 }
        multiplier: { type: number }
        disabled: { type: boolean }
    QuotaTable:
      type: object
      properties:
        unit: { type: string, example: 配额 }
        formula: { type: string, example: "消耗 = 端点系数(端点, 平台) × 条数 × 账号倍率" }
        endpoints: { type: object, additionalProperties: { type: string } }
        preauth_max:
          type: object
          additionalProperties: { type: number }
          description: 各端点的预授权上限（该端点在所有平台中的最高系数）
        platforms: { type: object, additionalProperties: true }
        note: { type: string }